DWORD and LONG are 32 bits off Windows, as they are in Win32
Arming the golden comparison found this on its first run: CArchive wrote 47
bytes where real MFC writes 39. Both `long` and `DWORD` were going out 8
bytes wide, because DWORD was spelled `unsigned long` and Unix is LP64.

afx.h documented the divergence deliberately - standard C++ has no second
32-bit integer type distinct from `int`, so DWORD could be 32 bits wide OR
distinct from UINT, not both, and distinctness was chosen to keep MFC's
overload sets well-formed. The comment called it "a trap waiting for the
eventual Linux port". The port arrived.

Width wins instead. Every consequence of the wrong width is silent: struct
layouts shift and serialized bytes change, and eMule's on-disk formats are
read by other clients, so a widened field is a corrupt file rather than an
internal detail. The cost turned out to be two overload pairs - CArchive's
UINT/DWORD in each direction - which collapse into one off Windows and are
#ifdef'd out there. Nothing else depended on the distinction. atlbase.h's
ULONG/LONG/REGSAM/LPDWORD get the same treatment.

CArchive's `long` slot stays 32 bits wide off Windows too, narrowing and
sign-extending across it: the wire format is not ours to redefine, a file
written here has to be readable by a real MFC build.

The Windows interface is untouched - every change is behind #ifdef _WIN32 -
so what eMule compiles against is unchanged.

Also commits the recording itself, so golden_conformance is now armed:
730 of 733 cases byte-identical to real MFC on this POSIX build.

+++ include/atlbase.h
@@ -30,12 +30,14 @@
 // eMule uses it for the shell/file-association settings; signatures
 // verified against the Microsoft Learn CRegKey class page.
 #ifndef _WIN32
+// All 32 bits wide, as in Win32 -- see the typedef block in afx.h for why
+// `unsigned long` is the wrong spelling of that off Windows.
 using HKEY = void*;
-using REGSAM = unsigned long;
+using REGSAM = unsigned int;            // Win32: ACCESS_MASK, i.e. a DWORD
 using LPSECURITY_ATTRIBUTES = void*;
-using LPDWORD = unsigned long*;
-using ULONG = unsigned long;
-using LONG = long;
+using LPDWORD = unsigned int*;          // Win32: DWORD*
+using ULONG = unsigned int;
+using LONG = int;
 #define REG_NONE 0
 #define REG_SZ 1
 #define REG_OPTION_NON_VOLATILE 0
